[CLIENT] Changed, that connections are reloaded _before_ getting a connection
Currently get_connections asks the connection collection for a
connection and then, under certain conditions, executes reload_connections
which can close the retrieved connection if the used transport implements
the __close_connections method.

This patch reorders the get_connection code so that the connection
retrieval happens as late as possible (after reload_connections).

Closes #292